Add CLI Implementation
