+++ build.gradle.kts
@@ -10,6 +10,7 @@ allprojects {
 plugins {
   id("org.jlleitschuh.gradle.ktlint") version "9.2.1"
   kotlin("jvm") version "1.3.72"
+  kotlin("plugin.serialization") version "1.3.72"
 }
 
 task("dockerUp", Exec::class) {
