+++ server/src/main/kotlin/me/cewong/smolurl/controllers/UrlApi.kt
@@ -10,8 +10,8 @@ import io.ktor.routing.route
 import io.ktor.util.pipeline.PipelineContext
 import java.io.File
 import java.io.FileNotFoundException
+import me.cewong.smolurl.model.UrlModel
 import me.cewong.smolurl.models.ErrorType
-import me.cewong.smolurl.models.UrlModel
 import me.cewong.smolurl.models.handleError
 import me.cewong.smolurl.services.Error
 import me.cewong.smolurl.services.Success
