+++ settings.gradle.kts
@@ -2,5 +2,6 @@ rootProject.name = "smolurl"
 
 include(
   ":cli",
+  ":model",
   ":server"
 )
